Move common.h to common folder
